requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==43.0.1
defusedxml==0.8.0rc2
Django==5.1.1
django-filter==24.3
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-yasg==1.21.7
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
PyJWT==2.9.0
pytest==8.3.3
pytest-django==4.9.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
simplejson==3.19.3
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
